Move Event Type enum from DockerClient to Event
